[#15] Move Background widget into widget package
